Remove outer class
