+++ backend/api/communication/src/test/java/co/airy/core/api/communication/ConversationsTagTest.java
@@ -2,7 +2,7 @@ package co.airy.core.api.communication;
 
 import co.airy.avro.communication.Channel;
 import co.airy.avro.communication.ChannelConnectionState;
-import co.airy.core.api.communication.util.ConversationGenerator;
+import co.airy.core.api.communication.util.TestConversation;
 import co.airy.kafka.schema.application.ApplicationCommunicationChannels;
 import co.airy.kafka.schema.application.ApplicationCommunicationMessages;
 import co.airy.kafka.schema.application.ApplicationCommunicationMetadata;
@@ -87,7 +87,7 @@ class ConversationsTagTest {
 
         final String conversationId = UUID.randomUUID().toString();
         kafkaTestHelper.produceRecords(
-                ConversationGenerator.TestConversation.from(
+                TestConversation.from(
                         conversationId,
                         channel,
                         1).getRecords());
