+++ backend/api/communication/src/test/java/co/airy/core/api/communication/UnreadCountTest.java
@@ -2,7 +2,7 @@ package co.airy.core.api.communication;
 
 import co.airy.avro.communication.Channel;
 import co.airy.avro.communication.ChannelConnectionState;
-import co.airy.core.api.communication.util.ConversationGenerator;
+import co.airy.core.api.communication.util.TestConversation;
 import co.airy.kafka.schema.application.ApplicationCommunicationChannels;
 import co.airy.kafka.schema.application.ApplicationCommunicationMessages;
 import co.airy.kafka.schema.application.ApplicationCommunicationMetadata;
@@ -88,7 +88,7 @@ class UnreadCountTest {
 
         final int unreadMessages = 3;
 
-        kafkaTestHelper.produceRecords(ConversationGenerator.TestConversation.from(
+        kafkaTestHelper.produceRecords(TestConversation.from(
                 conversationId,
                 channel,
                 unreadMessages).getRecords());
